+++ crates/adam-core/src/ids.rs
@@ -26,6 +26,8 @@ macro_rules! typed_id {
 }
 
 typed_id!(CountryId);
+typed_id!(GoodId);
+typed_id!(NeedProfileId);
 typed_id!(CohortId);
 typed_id!(RegionId);
 typed_id!(ActorId);
@@ -46,6 +48,8 @@ mod tests {
     #[test]
     fn ids_have_stable_compact_layout() {
         assert_eq!(size_of::<CountryId>(), size_of::<u32>());
+        assert_eq!(size_of::<GoodId>(), size_of::<u32>());
+        assert_eq!(size_of::<NeedProfileId>(), size_of::<u32>());
         assert_eq!(size_of::<CohortId>(), size_of::<u32>());
         assert_eq!(size_of::<RegionId>(), size_of::<u32>());
         assert_eq!(size_of::<ActorId>(), size_of::<u32>());